FIX: since TimeLeft it is initialized several times, only create the link the first time.
git-svn-id: svn+ssh://svn.cern.ch/reps/dirac/DIRAC/trunk/DIRAC@26753 4525493e-7705-40b1-a816-d608a930855b

+++ Core/Utilities/TimeLeft/LSFTimeLeft.py
@@ -112,7 +112,8 @@ class LSFTimeLeft:
 
       if not self.normRef:
         # Now parse LSF configuration files
-        os.symlink( os.path.join( os.environ['LSF_ENVDIR'], 'lsf.conf') , './lsf.sh' )
+        if not os.path.isfile( './lsf.sh' ):
+          os.symlink( os.path.join( os.environ['LSF_ENVDIR'], 'lsf.conf' ) , './lsf.sh' )
         ret = sourceEnv( 10, ['./lsf' ] )
         if ret['OK']:
           lsfEnv = ret['outputEnv']
